+++ routes/web.php
@@ -3,6 +3,8 @@
 use App\Http\Controllers\Frontend\HomeController;
 use App\Http\Controllers\Frontend\PredictionDetailController;
 use App\Http\Controllers\Frontend\PredictionsController;
+use App\Http\Controllers\Frontend\PromotionDetailController;
+use App\Http\Controllers\Frontend\PromotionsController;
 use App\Http\Controllers\Frontend\ResultDetailController;
 use App\Http\Controllers\Frontend\ResultsController;
 use Illuminate\Support\Facades\Route;
@@ -34,3 +36,10 @@ Route::get(
         'resultDate' => '\d{4}-\d{2}-\d{2}',
     ])
     ->name('results.show');
+
+Route::get('/promosi', PromotionsController::class)
+    ->name('promotions.index');
+
+Route::get('/promosi/{slug}', PromotionDetailController::class)
+    ->where('slug', '[a-z0-9]+(?:-[a-z0-9]+)*')
+    ->name('promotions.show');